Fix archetype icons not working.
They were referencing wrong place. I guess the subfolder is added in 1.9. This caused icons to not appear on Operation Menu, tho i think they are used at other places too, didn't check if they didn't work there either.

Also adds seperate one for AT Armored Car but it is a duplicate of normal one, for whenever someone decides to make a icon for it.