Fix "unique" to "ordered" add some names
The names of the destroyers are out of order. At the end are those that are in the fleet at the start. Also carrier names "Pegasopolis Class" so it ship name and other firs pony countries as "Unicornia" and "Dirtville" also add in this list
Maybe add some new destroyer names? 78 is not much for country like EQS