Speed League: Fuck the Order, all my homies hate the Order.
All of the stuff up to war and coring of the order is done.
Loc is mostly done (missing event loc.)
GFX is also mostly done (ideas and decisions don't have proper GFX)
But most importantly: ALL THE EFFECTS ARE DONE. (well, for that part)

Also all the decisions work.