Moving stuff from zzz_technologies.gfx to eaw_technologies.gfx.
zzz_technologies.gfx appears to be copied over from basegame Technologies.gfx and has a lot of basegame definitions for no reason. It is missing stuff from basegame that was added since and includes extra stuff that we added. I'm moving the latter, and if in the end nothing necessary is left we can delete the file. But i don't have time to finish it rn.