Sending Attachés Abroad focus and decision tree
A hefty one this focus is, but I did it. Changed some stuff in Gytha-Cenhelm and BEG formation to account for the new mechanic, fixed BEG's leader not being promoted sometimes. Added military leader training mechanic which allows to send a leader away for 5 months and buff them, only once tho. Added WAT and FIR military leader handlers for this system in case BEG integrates them.