AI production efficiency fix
This hidden modifier should help treat the constant swap of production lines for AI. Due to the swap, it loses a lot of equipment that the player would not have lost because of having brain. Therefore, it is only used starting from the regular difficulty.

Thanks WA mod for idea!

I made a few tests and pretty satisfied with results