Daring Do Quest Map Implementation
Modification to CES_decision_categories (adds GUI and ensures it is visible even if no decisions are being taken)

Addition of four decisions, three effects and six events needed for scripted GUI functionality. There are two continue decisions to avoid issues where a decision is called before it can properly disappear. Hidden state events are utilised to take advantage of the FROM scope.

Scripted localisation has been added for the decisions. If unique localisation is desired for the events, this same template can be followed.

The graphical assets needed for the mechanic have also been added.