Removing an error
I had a wrong log for a focus, as well as now also changing the modifier to be 11 so that whenever the focus tree changes or is loaded, it will load the new one over the old one. Preferably, this would use the availability of a country flag for this, but I don't know which one that is.