Removed 1 line of code.
Got around to fixing this decision after waiting 2 weeks. Tested it. Decision "Pursue detente with Equestria" NOW only appears if you are not in a faction w/ HIP + have completed 1 of 2 focuses, vs. before where if HIP was a faction leader it also wouldn't appear. For some reason. Still cannot take this decision if a puppet or in HIP's faction.