Move list of traits that enable ideology decisions to scripted triggers and make Feather Book replace PCB's generic democratic advisor.
Doesn't help much for EaW, we don't normally use these decisions with countries with content and generic countries don't use characters with custom traits for this. I wanted to make Feather Book in PCB replace their  generic Harmonic Reformer, but seems like he is currently only available while you are Harmonic and i didn't fell like removing that.

The real reason i needed this is for ESI-EaW so can add more traits here without overriding the whole political_decisions.txt file. But if someone decides to give a contentless county an ideology advisor with a custom trait, it can be added here to trigger the decision.